separate critical css and remix stuff on canvas and in designer
We need to reuse canvas stuff on the published site as well
