+++ apps/designer/app/shared/remix/root.tsx
@@ -1,9 +1,9 @@
 import { Links, LiveReload, Meta, Outlet, Scripts } from "@remix-run/react";
-import { CriticalCss } from "@webstudio-is/react-sdk";
 import { Env } from "~/shared/env";
 import { useThemeProps } from "~/shared/theme";
+import { CRITICAL_CSS_MARKER } from "./constants";
 
-export const Designer = () => {
+export const Root = () => {
   const themeProps = useThemeProps();
   return (
     <html lang="en" {...themeProps}>
@@ -12,7 +12,7 @@ export const Designer = () => {
         <meta name="viewport" content="width=device-width,initial-scale=1" />
         <Meta />
         <Links />
-        <CriticalCss />
+        {typeof document === "undefined" ? CRITICAL_CSS_MARKER : null}
       </head>
       <body>
         <Outlet />
