--- packages/react-sdk/src/canvas/index.ts
@@ -1,1 +1,0 @@
-export * from "./canvas";
