+++ packages/react-sdk/src/index.ts
@@ -6,4 +6,4 @@ export * as componentsMeta from "./components/meta";
 export * from "./user-props";
 export * from "./pubsub";
 export * from "./db";
-export * from "./canvas";
+export * from "./remix";
