+++ packages/react-sdk/src/remix/root.tsx
@@ -1,5 +1,5 @@
 import { Links, Meta, Outlet as DefaultOutlet } from "@remix-run/react";
-import { CriticalCss } from "..";
+import { CRITICAL_CSS_MARKER } from "./constants";
 
 /**
  * We are using Outlet prop from index layout when user renders site from a subdomain.
@@ -18,7 +18,7 @@ export const Root = ({
         <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon" />
         <Meta />
         <Links />
-        <CriticalCss />
+        {typeof document === "undefined" ? CRITICAL_CSS_MARKER : null}
       </head>
       <Outlet />
     </html>
