+++ packages/react-sdk/src/stitches/index.ts
@@ -1,3 +1,2 @@
 export * from "./to-css";
-export * from "./css";
-export * from "./critical-css";
+export * from "./stitches";
